- reworking the available configurations - no more non-Unicode ones - fixing a number of warning regarding signed/unsigned - adding CreateRotationMatrix to mat3d for convenience - extending exyz to split the same atom according to the number and model of the attached disorder - adding basis for polymeric Frameworks

+++ glib/glmaterial.cpp
@@ -229,7 +229,7 @@ void TGlMaterial::FromString(const olxstr& str, bool safe) {
   }
   TStrList toks(str, ';');
   Flags = toks[0].ToInt();
-  int ind = 1;
+  size_t ind = 1;
   if (!safe) {
     if ((Flags & sglmAmbientF) != 0) {
       AmbientF.FromString(toks[ind++]);
